refactor: profit 클래스 추가
